+++ bootstrap.php
@@ -4,7 +4,7 @@ use TitusPiJean\Flarum\Auth\SSOwat\Listener;
 use Illuminate\Contracts\Events\Dispatcher;
 
 return function (Dispatcher $events) {
-    $events->subscribe(Listener\AddClientAssets::class);
-    $events->subscribe(Listener\AddSSOwatAuthRoute::class);
-    $events->subscribe(Listener\AddApiAttributes::class);
+    $events->subscribe(Listeners\AddClientAssets::class);
+    $events->subscribe(Listeners\AddSSOwatAuthRoute::class);
+    $events->subscribe(Listeners\AddApiAttributes::class);
 };
