+++ src/Listeners/AddSSOwatAuthRoute.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TitusPiJean\Flarum\Auth\SSOwat\Listener;
+namespace TitusPiJean\Flarum\Auth\SSOwat\Listeners;
 
 use Flarum\Event\ConfigureForumRoutes;
 use Illuminate\Contracts\Events\Dispatcher;
@@ -20,6 +20,6 @@ class AddSSOwatAuthRoute
      */
     public function configureForumRoutes(ConfigureForumRoutes $event)
     {
-        $event->get('/auth/ssowat', 'auth.ssowat', 'TitusPiJean\Flarum\Auth\SSOwat\SSOwatAuthController');
+        $event->get('/auth/ssowat', 'auth.ssowat', 'TitusPiJean\Flarum\Auth\SSOwat\Controllers\SSOwatLogInController');
     }
 }
